Remove deprecated SocketUtils
Closes gh-28052
